enabled add whiskey as menu option, and whiskey list view works
