+++ WhiskyBarrl/app/src/main/java/edu/wpi/mis270xteam1/whiskybarrl/MainWhiskeyListActivity.java
@@ -1,7 +1,10 @@
 package edu.wpi.mis270xteam1.whiskybarrl;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
@@ -11,7 +14,6 @@ public class MainWhiskeyListActivity extends AppCompatActivity {
 
     private ListView whiskeyListView;
     private DatabaseHandler db;
-    private ArrayAdapter<Whiskey> whiskeyListAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -25,16 +27,40 @@ public class MainWhiskeyListActivity extends AppCompatActivity {
     }
 
     @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main_whiskey_list, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        if (id == R.id.menu_item_new_whiskey) {
+            // Start a new activity to add a whiskey
+            Intent i = new Intent(MainWhiskeyListActivity.this, NewWhiskeyActivity.class);
+            startActivity(i);
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
     protected void onResume() {
         super.onResume();
         loadWhiskeys();
     }
 
     private void loadWhiskeys() {
+        // Part of this taken from http://viralpatel.net/blogs/convert-arraylist-to-arrays-in-java/
         List<Whiskey> whiskeys = db.getAllWhiskeys();
         Whiskey[] whiskeyArray = whiskeys.toArray(new Whiskey[whiskeys.size()]);
 
-        whiskeyListAdapter = new WhiskeyListAdapter(this, whiskeyArray);
+        ArrayAdapter<Whiskey> whiskeyListAdapter = new WhiskeyListAdapter(this, whiskeyArray);
         whiskeyListView.setAdapter(whiskeyListAdapter);
     }
 }
