+++ WhiskyBarrl/app/src/main/res/layout/activity_main_whiskey_list.xml
@@ -14,6 +14,6 @@
         android:layout_height="wrap_content"
         android:id="@+id/whiskeyListView"
         android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true" />
+        android:layout_alignParentTop="true" />
 
 </RelativeLayout>
