+++ WhiskyBarrl/app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">WhiskyBarrl</string>
 
     <string name="action_settings">Settings</string>
+    <string name="new_whiskey">Add Whiskey</string>
 </resources>
